update type of some variables from size_t to int

+++ sort.h
@@ -4,7 +4,7 @@
 void selection_sort(int *array, size_t array_size);
 void bubble_sort(int *array, size_t array_size);
 void insertion_sort(int *array, size_t array_size);
-void merge_sort(int *array, size_t l, size_t r);
-void quick_sort(int *array, size_t l, size_t r);
+void merge_sort(int *array, int l, int r);
+void quick_sort(int *array, int l, int r);
 
 #endif
